Вынес отдельный файл Commands.cs, который хранит enum Commands
Создал новый класс Game, который управляет аквариумом

Создал новый класс GeneratorFishNames, который ответственный за генерацию имён

Fish:
Убрал генерацию имени как ответственность самой рыбы (Убрал метод GenerateFishName)
