+++ Aquarium/Aquarium.csproj
@@ -41,7 +41,10 @@
     </ItemGroup>
     <ItemGroup>
         <Compile Include="Aquarium.cs" />
+        <Compile Include="Commands.cs" />
         <Compile Include="Fish.cs" />
+        <Compile Include="Game.cs" />
+        <Compile Include="GeneratorFishNames.cs" />
         <Compile Include="Program.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="UserUtils.cs" />
