+++ Aquarium/Commands.cs
@@ -1,0 +1,10 @@
+namespace Aquarium
+{
+    public enum Commands : byte
+    {
+        AddFish = 1,
+        TakeOutFish,
+        Skip,
+        BrakeAquarium
+    }
+}
